Feat. 채팅 기능 구현

+++ Devlog/src/main/java/com/devlog/project/chatting/mapper/ChattingMapper.java
@@ -12,10 +12,10 @@ public interface ChattingMapper {
 	
 	
 	// 채팅방 목록 조회
-	List<ChattingListDTO> selectChatList(int memberNo);
+	List<ChattingListDTO> selectChatList(Long memberNo);
 	
 	
 	// 팔로우 회원 목록 조회
-	List<FollowListDTO> selectFollowList(int memberNo);
+	List<FollowListDTO> selectFollowList(Long memberNo);
 
 }
